Ahora si hicimos algo decente. lectura de temo y hum

+++ GccApplication3/Include/Header/DHT11.h
@@ -4,6 +4,6 @@
 #include <avr/io.h>
 
 void DHT11_Init(void);
-uint8_t DHT11_Read(uint8_t *temperature, uint8_t *humidity);
+uint8_t DHT11_Read(float *temperature, float *humidity);
 
 #endif /* DHT11_H_ */
